Moved multipart tests to plugin tests
